requirements.txt
Django==3.2.16
django-cors-headers==3.13.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-util==1.0.17
drf-yasg==1.21.4
executing==1.2.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
swagger-spec-validator==3.0.3
urllib3==1.26.16
django_extensions==3.2.3